+++ lapki/Views/AppointmentsView/AppointmentsView.swift
@@ -1,0 +1,20 @@
+//
+//  AppointmentsView.swift
+//  lapki
+//
+//  Created by Maxim Makarenkov on 11.11.2024.
+//
+
+import SwiftUI
+
+struct AppointmentsView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+            .navigationTitle("Мои записи")
+            .navigationBarTitleDisplayMode(.large)
+    }
+}
+
+#Preview {
+    AppointmentsView()
+}
