+++ lapki/Views/Layouts/PetItem.swift
@@ -1,0 +1,34 @@
+//
+//  PetItem.swift
+//  lapki
+//
+//  Created by Maxim Makarenkov on 06.11.2024.
+//
+
+import SwiftUI
+
+struct PetItem: View {
+    var body: some View {
+        NavigationLink(destination: PetView()) {
+            HStack {
+                Image(systemName: "cat.circle")
+                    .resizable()
+                    .scaledToFit()
+                    .frame(height: 40)
+                VStack(alignment: .leading) {
+                    Text("Барсик")
+                        .font(.headline)
+                    Text("4 года, дворняга")
+                        .font(.callout)
+                        .foregroundStyle(.gray)
+                }
+                Spacer()
+                Image(systemName: "chevron.right")
+            }
+        }
+    }
+}
+
+#Preview {
+    PetItem()
+}
