+++ lapki/Views/SettingsView/SettingsView.swift
@@ -1,0 +1,20 @@
+//
+//  SettingsView.swift
+//  lapki
+//
+//  Created by Maxim Makarenkov on 11.11.2024.
+//
+
+import SwiftUI
+
+struct SettingsView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+            .navigationTitle("Настройки")
+            .navigationBarTitleDisplayMode(.large)
+    }
+}
+
+#Preview {
+    SettingsView()
+}
