+++ project.yml
@@ -1,12 +1,6 @@
 name: lapki
 options:
   bundleIdPrefix: ru.inTeam
-settings:
-  base:
-    SUPPORTS_MAC_DESIGNED_FOR_IPHONE_IPAD: NO
-    SUPPORTS_XR_DESIGNED_FOR_IPHONE_IPAD: NO
-    TARGETED_DEVICE_FAMILY: "1"
-    PRODUCT_BUNDLE_IDENTIFIER: $(BUNDLE_ID_PREFIX).$(TARGET_NAME)
 targets:
   lapki:
     type: application
@@ -28,3 +22,8 @@ targets:
             UIWindowSceneSessionRoleApplication:
               - UISceneConfigurationName: "Default Configuration"
                 UISceneDelegateClassName: "$(PRODUCT_MODULE_NAME).SceneDelegate"
+    settings:
+      base:
+        SUPPORTS_MAC_DESIGNED_FOR_IPHONE_IPAD: NO
+        SUPPORTS_XR_DESIGNED_FOR_IPHONE_IPAD: NO
+        TARGETED_DEVICE_FAMILY: "1"
